MappedTransferContainerTest moved to ContainerTest transfer package
