--- judo-runtime-core-jsl-itest/models/MappedTransferContainerTestModel/pom.xml
@@ -1,64 +1,0 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-
-    <parent>
-        <groupId>hu.blackbelt.judo.runtime</groupId>
-        <artifactId>judo-runtime-core-jsl-itest-model</artifactId>
-        <version>${revision}</version>
-    </parent>
-
-    <artifactId>judo-runtime-core-jsl-itest-model-MappedTransferContainerTestModel</artifactId>
-    <packaging>jar</packaging>
-
-    <name>JUDO Runtime Core :: JUDO Language Specification DSL Integration Tests :: MappedTransferContainerTestModel</name>
-
-    <build>
-        <plugins>
-            <plugin>
-                <groupId>hu.blackbelt.judo.tatami</groupId>
-                <artifactId>judo-tatami-jsl-workflow-maven-plugin</artifactId>
-                <version>${judo-tatami-jsl-version}</version>
-            </plugin>
-
-            <plugin>
-                <groupId>hu.blackbelt.judo.meta</groupId>
-                <artifactId>judo-psm-generator-maven-plugin</artifactId>
-                <version>${judo-meta-psm-version}</version>
-                <executions>
-                    <execution>
-                        <id>execute-psm-test-model-from-artifact-MappedTransferContainerTestModel</id>
-                        <phase>generate-test-sources</phase>
-                        <goals>
-                            <goal>generate</goal>
-                        </goals>
-                        <configuration>
-                            <psm>
-                                ${basedir}/target/generated-test-sources/model/MappedTransferContainerTest-psm.model
-                            </psm>
-                            <destination>${basedir}/target/generated-test-sources/sdk-core/MappedTransferContainerTestModel</destination>
-                        </configuration>
-                    </execution>
-                </executions>
-            </plugin>
-
-            <plugin>
-                <groupId>org.codehaus.mojo</groupId>
-                <artifactId>build-helper-maven-plugin</artifactId>
-                <executions>
-                    <execution>
-                        <id>add-test-source</id>
-                        <phase>generate-test-sources</phase>
-                        <goals>
-                            <goal>add-test-source</goal>
-                        </goals>
-                        <configuration>
-                            <sources>
-                                <source>${project.basedir}/target/generated-test-sources/sdk-core/MappedTransferContainerTestModel</source>
-                            </sources>
-                        </configuration>
-                    </execution>
-                </executions>
-            </plugin>
-        </plugins>
-    </build>
-</project>
