--- judo-runtime-core-jsl-itest/models/MappedTransferContainerTestModel/src/test/resources/MappedTransferContainerTest.jsl
@@ -1,47 +1,0 @@
-model MappedTransferContainerTest;
-
-type string String min-size:0 max-size:255;
-
-entity A {
-    field required C conA;
-}
-
-entity B extends A {
-    field required D donB;
-}
-
-entity C {
-    relation A containerA <= self!container(entityType = A);
-    relation B containerB <= self!container(entityType = A)!asType(entityType = B);
-    relation B containerB1 <= self!container(entityType = B);
-}
-
-entity D {
-    relation A containerA <= self!container(entityType = B)!asType(entityType = A);
-    relation A containerA1 <= self!container(entityType = A);
-    relation B containerB <= self!container(entityType = B);
-}
-
-transfer TA maps A as a {
-    @Embedded
-    relation required TC conA <=> a.conA;
-}
-
-transfer TB maps B as b {
-    @Embedded
-    relation required TD donB <=> b.donB;
-    @Embedded
-    relation required TC conA <=> b.conA;
-}
-
-transfer TC maps C as c {
-    relation TA containerA <= c.containerA;
-    relation TB containerB <= c.containerB;
-    relation TB containerB1 <= c.containerB1;
-}
-
-transfer TD maps D as d {
-    relation TA containerA <= d.containerA;
-    relation TA containerA1 <= d.containerA1;
-    relation TB containerB <= d.containerB;
-}
