+++ judo-runtime-core-jsl-itest/models/pom.xml
@@ -35,7 +35,6 @@
         <module>MappedTransferCastingFunctionsModel</module>
         <module>MappedTransferCompositonAggregationModel</module>
         <module>MappedTransferCompositonAssociationModel</module>
-        <module>MappedTransferContainerTestModel</module>
         <module>MappedTransferContainsTestModel</module>
         <module>MappedTransferFilterModel</module>
         <module>MappedTransferFunctionsModels</module>
